+++ apps/playground/src/app/components/lookup/selection-modal/selection-modal.module.ts
@@ -1,0 +1,13 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+import { SelectionModalRoutingModule } from './selection-modal-routing.module';
+import { SelectionModalComponent } from './selection-modal.component';
+
+@NgModule({
+  declarations: [SelectionModalComponent],
+  imports: [SelectionModalRoutingModule, CommonModule],
+})
+export class SelectionModalModule {
+  public static routes = SelectionModalRoutingModule.routes;
+}