+++ libs/components/datetime/src/lib/modules/timepicker/fixtures/timepicker-input-box-component.fixture.ts
@@ -16,6 +16,8 @@ export class TimepickerInputBoxTestComponent {
   @ViewChild(SkyTimepickerInputDirective)
   public timepicker!: SkyTimepickerInputDirective;
 
+  public labelText = 'Time';
+
   public returnFormat: string | undefined;
 
   public selectedTime: SkyTimepickerTimeOutput | undefined;